+++ src/design/plone/contenttypes/adapters/searchabletext_indexers.py
@@ -2,15 +2,19 @@
 from collective.dexteritytextindexer.converters import (
     DefaultDexterityTextIndexFieldConverter,
 )
-from collective.dexteritytextindexer.interfaces import (
-    IDexterityTextIndexFieldConverter,
-)
+from collective.dexteritytextindexer.interfaces import IDexterityTextIndexFieldConverter
 from plone.dexterity.interfaces import IDexterityContent
+from plone.restapi.interfaces import IBlockSearchableText
 from z3c.form.interfaces import IWidget
 from z3c.relationfield.interfaces import IRelationChoice
 from z3c.relationfield.interfaces import IRelationList
 from zope.component import adapter
 from zope.interface import implementer
+from zope.publisher.interfaces.browser import IBrowserRequest
+from design.plone.contenttypes.interfaces import IDesignPloneContentType
+from plone.restapi.indexers import (
+    TextBlockSearchableText as BaseTextBlockSearchableText,
+)
 
 
 @implementer(IDexterityTextIndexFieldConverter)
@@ -33,6 +37,10 @@ class RelationListFieldConverter(DefaultDexterityTextIndexFieldConverter):
         relations = self.field.get(self.context)
         if not relations:
             return ""
-        return " ".join(
-            [x.to_object.Title() for x in relations if x.to_object]
-        )
+        return " ".join([x.to_object.Title() for x in relations if x.to_object])
+
+
+@implementer(IBlockSearchableText)
+@adapter(IDesignPloneContentType, IBrowserRequest)
+class TextBlockSearchableText(BaseTextBlockSearchableText):
+    """ """