+++ src/design/plone/contenttypes/interfaces/modulo.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
+from design.plone.contenttypes.interfaces import IDesignPloneContentType
 from plone.supermodel import model
 
 
-class IModulo(model.Schema):
+class IModulo(model.Schema, IDesignPloneContentType):
     """ Modulo """